\name{LogisticRCI-package}
\alias{LogisticRCI-package}
\alias{LogisticRCI}
\encoding{UTF-8}
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\docType{package}
\title{
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\packageTitle{LogisticRCI}
}
\description{
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\packageDescription{LogisticRCI}
}
\details{

\strong{Linear and Logistic Regression-Based Reliable Change Index}

Here we provide an implementation of the linear and logistic regression-based Reliable Change Index (RCI), to be used with lm and binomial glm model objects, respectively. The \code{RCI} function returns a score assumed to be approximately normally distributed, which is helpful to detect patients that may present cognitive decline.

}

\author{
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
\packageAuthor{LogisticRCI}

%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
Maintainer: \packageMaintainer{LogisticRCI}
}
\references{
Moral, R.A., Diaz-Orueta, U., Oltra-Cucarella, J. (preprint) Logistic versus linear regression-based Reliable Change Index: implications for clinical studies with diverse sample sizes. DOI: 10.31234/osf.io/gq7az
}
\keyword{ package }
\seealso{
\code{\link{RCI}}
}